Refactoring and updates to rate limiting
Move rate limit code and parsing code to separate files
Make rate limiting list structure more efficient
