+++ usr.sbin/inetd/inetd.h
@@ -66,11 +66,13 @@
 #include <sys/socket.h>
 #include <sys/time.h>
 #include <sys/un.h>
+#include <sys/queue.h>
 
 #include <arpa/inet.h>
 
 #include <netdb.h>
 #include <stdbool.h>
+
 
 #include "pathnames.h"
 
@@ -121,6 +123,21 @@ typedef enum service_type {
 #define STRINGIFY(x) #x
 #define TOSTRING(x) STRINGIFY(x)
 
+/* "Unspecified" indicator value for servtabs (mainly used by v2 syntax) */
+#define SERVTAB_UNSPEC_VAL -1
+
+#define SERVTAB_UNSPEC_SIZE_T SIZE_MAX
+
+#define SERVTAB_COUNT_MAX (SIZE_MAX - (size_t)1)
+
+/* Standard logging and debug print format for a servtab */
+#define SERV_FMT "%s/%s"
+#define SERV_PARAMS(sep) sep->se_service,sep->se_proto
+
+/* rate limiting macros */
+#define	CNT_INTVL	((time_t)60)	/* servers in CNT_INTVL sec. */
+#define	RETRYTIME	(60*10)		/* retry after bind or server fail */
+
 struct	servtab {
 	char	*se_hostaddr;		/* host address to listen on */
 	char	*se_service;		/* name of service */
@@ -148,30 +165,79 @@ struct	servtab {
 	int	se_fd;			/* open descriptor */
 	service_type	se_type;	/* type */
 	union {
-		struct sockaddr_storage	se_ctrladdr_storage; /* ensure correctness of C struct initializer */
+		/* ensure correctness of C struct initializer */
+		struct sockaddr_storage	se_ctrladdr_storage;
 		struct sockaddr	se_ctrladdr;
 		struct sockaddr_in	se_ctrladdr_in;
 		struct sockaddr_in6	se_ctrladdr_in6; /* in6 is used by bind()/getaddrinfo */
 		struct sockaddr_un	se_ctrladdr_un;
 	};				/* bound address */
 	socklen_t	se_ctrladdr_size;
-	size_t	se_service_max;		/* max # of instances of this service */
+	size_t	se_service_max;		/* max # of instances of this service per minute */
 	size_t	se_count;		/* number of instances of this service started since se_time */
 	size_t	se_ip_max;  		/* max # of instances of this service per ip per minute */
-	struct se_ip_list_node {
-		struct se_ip_list_node	*next;
-		size_t count;		/*
-					 * number of instances of this service started from
-					 * this ip address since se_time (includes
-					 * attempted starts if greater than se_ip_max)
-					 */
-		char address[NI_MAXHOST];
-	} *se_ip_list_head; 		/* linked list of number of requests per ip */
+	SLIST_HEAD(iplist, rl_ip_node) se_rl_ip_list; /* per-address (IP) rate limting */
 	time_t se_time;	/* start of se_count and ip_max counts, in seconds from arbitrary point */
+	
+	/* TODO convert to using SLIST */
 	struct	servtab *se_next;
 };
 
-/* From inetd.c */
+struct rl_ip_node {
+	/* Linked list entries */
+	SLIST_ENTRY(rl_ip_node) entries;
+	/*
+	 * Number of service spawns from *_addr since se_time (includes
+	 * attempted starts if greater than se_ip_max).
+	 */
+	size_t count;
+	union {
+		struct in_addr	ipv4_addr;
+#ifdef INET6
+		/* align for efficient comparison in rl_try_get, could use 8 instead */
+		struct in6_addr	ipv6_addr __attribute__((aligned(16)));
+#endif
+		/* 
+		 * other_addr is used for other address types besides the 
+		 * special cases (IPv4/IPv6), using getnameinfo.
+		 */
+		struct {
+			/* A field is required before the special array member */
+			char _placeholder;
+			/* malloc'd storage varies with length of string */
+			char other_addr[];
+		};
+	};
+	/* 
+	 * Do not declare further members after union, offsetof is used to
+	 * determine malloc size.
+	 */
+};
+
+/* 
+ * From inetd.c
+ */
+
+void	setup(struct servtab *);
+void	close_sep(struct servtab *);
+void	register_rpc(struct servtab *);
+void	unregister_rpc(struct servtab *);
+bool	try_biltin(struct servtab *);
+
+/* Global debug mode boolean, enabled with -d */
+extern int debug;
+
+/* rate limit or other error timed out flag */
+extern int	timingout;
+
+/* servtab linked list */
+extern struct servtab *servtab;
+
+/*
+ * From parse.c
+ */
+
+void	config_root(void);
 int 	parse_protocol(struct servtab *);
 int 	parse_wait(struct servtab *, int);
 int 	parse_server(struct servtab *, const char *);
@@ -179,27 +245,32 @@ void 	parse_socktype(char *, struct servtab *);
 void 	parse_accept_filter(char *, struct servtab *);
 char 	*nextline(FILE *);
 char 	*newstr(const char *);
-void	freeconfig(struct servtab *);
 
-/* Global debug mode boolean, enabled with -d */
-extern int debug;
+/* Current line number in current config file */
+extern size_t	line_number;
 
 /* Current config file path */
-extern const char *CONFIG;
+extern const char	*CONFIG;
 
 /* Open config file */
 extern FILE	*fconfig;
 
-/* Current line number in current config file */
-extern size_t line_number;
-
 /* Default listening hostname/IP for current config file */
-extern char *defhost;
+extern char	*defhost;
 
 /* Default IPsec policy for current config file */
-extern char *policy;
+extern char	*policy;
 
-/* From parse_v2.c */
+/*
+ * From ratelimit.c 
+ */
+
+int	rl_process(struct servtab *, int);
+void	rl_clear_ip_list(struct servtab *);
+
+/*
+ * From parse_v2.c
+ */
 
 typedef enum parse_v2_result {V2_SUCCESS, V2_SKIP, V2_ERROR} parse_v2_result;
 
@@ -209,16 +280,5 @@ typedef enum parse_v2_result {V2_SUCCESS, V2_SKIP, V2_ERROR} parse_v2_result;
  * Fills the provided servtab structure. Does not call freeconfig on error.
  */
 parse_v2_result	parse_syntax_v2(struct servtab *, char **);
-
-/* "Unspecified" indicator value for servtabs (mainly used by v2 syntax) */
-#define SERVTAB_UNSPEC_VAL -1
-
-#define SERVTAB_UNSPEC_SIZE_T SIZE_MAX
-
-#define SERVTAB_COUNT_MAX (SIZE_MAX - (size_t)1)
-
-/* Standard logging and debug print format for a servtab */
-#define SERV_FMT "%s/%s"
-#define SERV_PARAMS(sep) sep->se_service,sep->se_proto
 
 #endif
